add helpers to utils lib
