+++ packages/ethers/src/utils/types.ts
@@ -47,11 +47,3 @@ export type Metadata = {
   url: string
   icons: string[]
 }
-
-export type Chain = {
-  rpcUrl: string
-  explorerUrl: string
-  currency: string
-  name: string
-  chainId: number
-}
