+++ packages/scaffold-utils/src/ethers/index.ts
@@ -1,1 +1,3 @@
 export { WALLET_ID, ERROR_CODE_UNRECOGNIZED_CHAIN_ID, ERROR_CODE_DEFAULT } from './constants.js'
+export { getCaipDefaultChain, hexStringToNumber, numberToHexString } from './helpers.js'
+export type { Chain } from './types.js'
