+++ packages/scaffold-utils/src/ethers/types.ts
@@ -1,0 +1,7 @@
+export type Chain = {
+  rpcUrl: string
+  explorerUrl: string
+  currency: string
+  name: string
+  chainId: number
+}
